Eliminate use of __tu::tuple except for agency::tuple Move several items from tuple_utility.hpp into tuple.hpp Eliminate some unused items from tuple_utility.hpp
