+++ agency/tuple/detail/tuple_utility.hpp
@@ -118,67 +118,6 @@ auto forward_tail(Tuple&& t)
 }
 
 
-struct tuple_mover
-{
-  template<class... Args>
-  __AGENCY_ANNOTATION
-  auto operator()(Args&&... args)
-    -> decltype(
-         __tu::make_tuple(std::move(args)...)
-       )
-  {
-    return __tu::make_tuple(std::move(args)...);
-  }
-};
-
-
-template<class Tuple>
-__AGENCY_ANNOTATION
-auto move_tail(Tuple&& t)
-  -> decltype(
-       __tu::tuple_tail_invoke(std::move(t), tuple_mover{})
-     )
-{
-  return __tu::tuple_tail_invoke(std::move(t), tuple_mover{});
-}
-
-
-struct agency_tuple_maker
-{
-  template<class... Args>
-  __AGENCY_ANNOTATION
-  auto operator()(Args&&... args)
-    -> decltype(
-         __tu::make_tuple(std::forward<Args>(args)...)
-       )
-  {
-    return __tu::make_tuple(std::forward<Args>(args)...);
-  }
-};
-
-
-template<class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_tail(Tuple&& t)
-  -> decltype(
-       __tu::tuple_tail_invoke(std::forward<Tuple>(t), agency_tuple_maker{})
-     )
-{
-  return __tu::tuple_tail_invoke(std::forward<Tuple>(t), agency_tuple_maker{});
-}
-
-
-template<class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_prefix(Tuple&& t)
-  -> decltype(
-       __tu::tuple_prefix_invoke(std::forward<Tuple>(t), agency_tuple_maker{})
-     )
-{
-  return __tu::tuple_prefix_invoke(std::forward<Tuple>(t), agency_tuple_maker{});
-}
-
-
 template<class T,
          class = typename std::enable_if<
            is_tuple<typename std::decay<T>::type>::value
@@ -207,52 +146,6 @@ template<class T,
            is_tuple<typename std::decay<T>::type>::value
          >::type>
 __AGENCY_ANNOTATION
-auto tuple_tail_if(T&& t) ->
-  decltype(detail::tuple_tail(std::forward<T>(t)))
-{
-  return detail::tuple_tail(std::forward<T>(t));
-}
-
-
-template<class T,
-         class = typename std::enable_if<
-           !is_tuple<typename std::decay<T>::type>::value
-         >::type>
-__AGENCY_ANNOTATION
-__tu::tuple<> tuple_tail_if(T&&)
-{
-  return __tu::tuple<>();
-}
-
-
-template<class T,
-         class = typename std::enable_if<
-           is_tuple<typename std::decay<T>::type>::value
-         >::type>
-__AGENCY_ANNOTATION
-auto tuple_prefix_if(T&& t) ->
-  decltype(detail::tuple_prefix(std::forward<T>(t)))
-{
-  return detail::tuple_prefix(std::forward<T>(t));
-}
-
-
-template<class T,
-         class = typename std::enable_if<
-           !is_tuple<typename std::decay<T>::type>::value
-         >::type>
-__AGENCY_ANNOTATION
-__tu::tuple<> tuple_prefix_if(T&&)
-{
-  return __tu::tuple<>();
-}
-
-
-template<class T,
-         class = typename std::enable_if<
-           is_tuple<typename std::decay<T>::type>::value
-         >::type>
-__AGENCY_ANNOTATION
 auto tuple_last_if(T&& t) ->
   decltype(__tu::tuple_last(std::forward<T>(t)))
 {
@@ -271,28 +164,6 @@ T&& tuple_last_if(T&& t)
 }
 
 
-template<typename Function, typename Tuple, typename... Tuples>
-__AGENCY_ANNOTATION
-auto tuple_map(Function f, Tuple&& t, Tuples&&... ts)
-  -> decltype(
-       __tu::tuple_map_with_make(f, agency_tuple_maker(), std::forward<Tuple>(t), std::forward<Tuples>(ts)...)
-     )
-{
-  return __tu::tuple_map_with_make(f, agency_tuple_maker(), std::forward<Tuple>(t), std::forward<Tuples>(ts)...);
-}
-
-
-template<size_t N, class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_take(Tuple&& t)
-  -> decltype(
-       __tu::tuple_take_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_take_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker());
-}
-
-
 template<size_t N, class Tuple>
 __AGENCY_ANNOTATION
 auto tuple_take_view(Tuple&& t)
@@ -301,17 +172,6 @@ auto tuple_take_view(Tuple&& t)
      )
 {
   return __tu::tuple_take_invoke<N>(std::forward<Tuple>(t), forwarder());
-}
-
-
-template<size_t N, class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_drop(Tuple&& t)
-  -> decltype(
-       __tu::tuple_drop_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_drop_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker());
 }
 
 
@@ -326,28 +186,6 @@ auto tuple_drop_view(Tuple&& t)
 }
 
 
-template<size_t N, class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_drop_back(Tuple&& t)
-  -> decltype(
-       __tu::tuple_drop_back_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_drop_back_invoke<N>(std::forward<Tuple>(t), agency_tuple_maker());
-}
-
-
-template<class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_drop_last(Tuple&& t)
-  -> decltype(
-       agency::detail::tuple_drop_back<1>(std::forward<Tuple>(t))
-     )
-{
-  return agency::detail::tuple_drop_back<1>(std::forward<Tuple>(t));
-}
-
-
 template<class Function, class Tuple>
 __AGENCY_ANNOTATION
 auto tuple_apply(Function&& f, Tuple&& t)
@@ -359,28 +197,6 @@ auto tuple_apply(Function&& f, Tuple&& t)
 }
 
 
-template<size_t N, class T>
-__AGENCY_ANNOTATION
-auto tuple_repeat(const T& x)
-  -> decltype(
-       __tu::tuple_repeat_invoke<N>(x, agency_tuple_maker())
-     )
-{
-  return __tu::tuple_repeat_invoke<N>(x, agency_tuple_maker());
-}
-
-
-template<template<class T> class MetaFunction, class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_filter(Tuple&& t)
-  -> decltype(
-       __tu::tuple_filter_invoke<MetaFunction>(std::forward<Tuple>(t), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_filter_invoke<MetaFunction>(std::forward<Tuple>(t), agency_tuple_maker());
-}
-
-
 template<template<class T> class MetaFunction, class Tuple>
 __AGENCY_ANNOTATION
 auto tuple_filter_view(Tuple&& t)
@@ -389,17 +205,6 @@ auto tuple_filter_view(Tuple&& t)
      )
 {
   return __tu::tuple_filter_invoke<MetaFunction>(std::forward<Tuple>(t), forwarder());
-}
-
-
-template<size_t... Indices, class Tuple>
-__AGENCY_ANNOTATION
-auto tuple_gather(Tuple&& t)
-  -> decltype(
-       __tu::tuple_gather_invoke<Indices...>(std::forward<Tuple>(t), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_gather_invoke<Indices...>(std::forward<Tuple>(t), agency_tuple_maker());
 }
 
 
@@ -462,43 +267,6 @@ template<class Tuple>
 using tuple_elements = typename tuple_elements_impl<tuple_indices<Tuple>,Tuple>::type;
 
 
-template<class Tuple, class T>
-__AGENCY_ANNOTATION
-auto tuple_append(Tuple&& t, T&& val)
-  -> decltype(
-       __tu::tuple_append_invoke(std::forward<Tuple>(t), std::forward<T>(val), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_append_invoke(std::forward<Tuple>(t), std::forward<T>(val), agency_tuple_maker());
-}
-
-
-template<class Tuple, class T>
-__AGENCY_ANNOTATION
-auto tuple_prepend(Tuple&& t, T&& val)
-  -> decltype(
-       __tu::tuple_prepend_invoke(std::forward<Tuple>(t), std::forward<T>(val), agency_tuple_maker())
-     )
-{
-  return __tu::tuple_prepend_invoke(std::forward<Tuple>(t), std::forward<T>(val), agency_tuple_maker());
-}
-
-
-template<class Tuple, class T>
-struct tuple_prepend_result
-{
-  using type = decltype(
-    detail::tuple_prepend(
-      std::declval<Tuple>(),
-      std::declval<T>()
-    )
-  );
-};
-
-template<class Tuple, class T>
-using tuple_prepend_result_t = typename tuple_prepend_result<Tuple,T>::type;
-
-
 template<class Tuple,
          class = typename std::enable_if<
            (std::tuple_size<
@@ -554,91 +322,6 @@ T&& unwrap_single_element_tuple_if(T&& arg)
 {
   return std::forward<T>(arg);
 }
-
-
-template<class TupleReference, class IndexSequence>
-struct decay_tuple_impl;
-
-
-template<class TupleReference, size_t... Indices>
-struct decay_tuple_impl<TupleReference, index_sequence<Indices...>>
-{
-  using tuple_type = typename std::decay<TupleReference>::type;
-
-  using type = __tu::tuple<
-    typename std::decay<
-      typename std::tuple_element<
-        Indices,
-        tuple_type
-      >::type
-    >::type...
-  >;
-};
-
-
-template<class TupleReference>
-struct decay_tuple : decay_tuple_impl<TupleReference, tuple_indices<typename std::decay<TupleReference>::type>> {};
-
-template<class TupleReference>
-using decay_tuple_t = typename decay_tuple<TupleReference>::type;
-
-
-template<class TypeList>
-struct homogeneous_tuple_impl;
-
-template<class... Types>
-struct homogeneous_tuple_impl<type_list<Types...>>
-{
-  using type = __tu::tuple<Types...>;
-};
-
-template<class T, size_t size>
-using homogeneous_tuple = typename homogeneous_tuple_impl<type_list_repeat<size,T>>:: type;
-
-
-template<size_t size, class T>
-__AGENCY_ANNOTATION
-homogeneous_tuple<T,size> make_homogeneous_tuple(const T& val)
-{
-  return detail::tuple_repeat<size>(val);
-}
-
-
-// this is the inverse operation of tuple_elements
-template<class TypeList>
-struct tuple_from_type_list;
-
-template<class... Types>
-struct tuple_from_type_list<agency::detail::type_list<Types...>>
-{
-  using type = __tu::tuple<Types...>;
-};
-
-template<class TypeList>
-using tuple_from_type_list_t = typename tuple_from_type_list<TypeList>::type;
-
-
-template<class TypeList>
-struct tuple_or_single_type_or_void_from_type_list
-{
-  using type = tuple_from_type_list_t<TypeList>;
-};
-
-template<class T>
-struct tuple_or_single_type_or_void_from_type_list<agency::detail::type_list<T>>
-{
-  using type = T;
-};
-
-template<>
-struct tuple_or_single_type_or_void_from_type_list<agency::detail::type_list<>>
-{
-  using type = void;
-};
-
-
-template<class TypeList>
-using tuple_or_single_type_or_void_from_type_list_t = typename tuple_or_single_type_or_void_from_type_list<TypeList>::type;
 
 
 template<class Indices, class Tuple>
@@ -702,45 +385,6 @@ struct is_empty_tuple_impl<Tuple, typename std::enable_if<is_tuple<Tuple>::value
 
 template<class Tuple>
 struct is_empty_tuple : is_empty_tuple_impl<Tuple>::type {};
-
-
-template<class Tuple>
-using tuple_reverse_t = tuple_from_type_list_t<
-  type_list_reverse<
-    tuple_elements<Tuple>
-  >
->;
-
-
-template<size_t n,
-         class T,
-         class = typename std::enable_if<
-           is_tuple<typename std::decay<T>::type>::value
-         >::type,
-         class = typename std::enable_if<
-           (n <= std::tuple_size<typename std::decay<T>::type>::value)
-         >::type>
-__AGENCY_ANNOTATION
-auto tuple_take_if(T&& t) ->
-  decltype(detail::tuple_take<n>(std::forward<T>(t)))
-{
-  return detail::tuple_take<n>(std::forward<T>(t));
-}
-
-
-template<size_t n,
-         class T,
-         class = typename std::enable_if<
-           !is_tuple<typename std::decay<T>::type>::value
-         >::type,
-         class = typename std::enable_if<
-           n == 1
-         >::type>
-__AGENCY_ANNOTATION
-typename std::decay<T>::type tuple_take_if(T&& value)
-{
-  return std::forward<T>(value);
-}
 
 
 // tuple_rebind takes a Tuple-like type and reinstantiates it with a different list of types
